+++ src/controllers/user/index.ts
@@ -1,0 +1,4 @@
+import userRouter from "./router"
+
+
+export default userRouter;
