+++ src/server.ts
@@ -1,6 +1,7 @@
 import app from './app'
 
-const PORT = 3000;
+const PORT = 4000;
 app.listen(PORT, () => {
+  console.log('Abhishek')
   console.log(`Server is running on port ${PORT}`);
 });
