+++ tsconfig.json
@@ -10,7 +10,7 @@
   },
   "include": [
     "src/**/*"
-  ],
+, "public/Day1"  ],
   "exclude": [
     "node_modules"
   ]
